requirements.txt
# Core typing and utility libraries
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
sniffio==1.3.1
tenacity==9.1.2
rsa==4.9.1

# Google + networking dependencies
google-auth==2.41.1
google-genai==1.44.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
requests==2.32.5
websockets==15.0.1

# Data validation
pydantic==2.12.2
pydantic_core==2.41.4
pyasn1==0.6.1
pyasn1_modules==0.4.2

# Web framework and helpers
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Pillow==10.1.0

# WSGI server (for production use)
waitress==2.1.2
